composer.json : 1-install google2fa package 2-install web push package 3-remove all pest package dockerFile: 1-change php -v from 8.2 to 8.3 2- install bcmath package in the php controllers: 1-remove AuthController.php 2-add Auth folder to controller folder 3-make AuthController separate into many smaller controller 4-createAccountController 5-move all controller to api folder 6-create web push controller webPush: 1-publish web push config file 2-add web push migration 3-add hasPushSubscription to user model 2FA: 1-add 2fa migration 2-create require 2fa middleware password reset: 1-implement password reset logic in controller

+++ app/Models/User.php
@@ -9,11 +9,12 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Sanctum\HasApiTokens;
+use NotificationChannels\WebPush\HasPushSubscriptions;
 
 class User extends Authenticatable implements MustVerifyEmail
 {
     /** @use HasFactory<\Database\Factories\UserFactory> */
-    use HasFactory, Notifiable, HasApiTokens;
+    use HasFactory, Notifiable, HasPushSubscriptions;
 
     /**
      * The attributes that are mass assignable.
@@ -53,8 +54,9 @@ class User extends Authenticatable implements MustVerifyEmail
 
     protected static function booted(){
         static::deleting(function ($user) {
-            $user->socialAccounts()->delete();
-            $user->image()->delete();
+            if($user->image){
+                $user->image()->delete();
+            }
         });
     }
